move detect_license funciton
