+++ thoth/license_solver/__init__.py
@@ -17,12 +17,36 @@
 
 """Init package."""
 
-from .solver import detect_license
+from .solver import Solver
+from typing import Dict, Any, List, Union
 
 
 __title__ = "license-solver"
 __version__ = "0.1.0"
 __author__ = "Viliam Podhajecky <vpodhaje@redhat.com>"
+
+
+def detect_license(
+    input_data: Union[Dict[str, Any], str, List[str], List[Dict[str, Any]]], raise_on_error: bool = True
+) -> Dict[str, Any]:
+    """Run license-solver from thoth-solver."""
+    try:
+        license_solver = Solver()
+
+        if type(input_data) == dict or type(input_data) == str:
+            license_solver.solve_from_file(input_data)
+        elif type(input_data) == list:
+            for enter in input_data:
+                license_solver.solve_from_file(enter)
+
+        return license_solver.get_output_dict()
+
+    except Exception:
+        if raise_on_error:
+            raise Exception
+
+        return dict()
+
 
 __all__ = [
     "detect_license",
