+++ thoth/license_solver/solver.py
@@ -34,28 +34,6 @@ from .exceptions import UnableOpenFileData
 _LOGGER = logging.getLogger(__name__)
 
 
-def detect_license(
-    input_data: Union[Dict[str, Any], str, List[str], List[Dict[str, Any]]], raise_on_error: bool = True
-) -> Dict[str, Any]:
-    """Run license-solver from thoth-solver."""
-    try:
-        solver = Solver()
-
-        if type(input_data) == dict or type(input_data) == str:
-            solver.solve_from_file(input_data)
-        elif type(input_data) == list:
-            for enter in input_data:
-                solver.solve_from_file(enter)
-
-        return solver.get_output_dict()
-
-    except Exception:
-        if raise_on_error:
-            raise Exception
-
-        return dict()
-
-
 @attr.s(slots=True)
 class Solver:
     """Class pass all detected files and try to detect all necessary data."""
@@ -212,6 +190,10 @@ class Solver:
         """Print final output on STDOU."""
         self.output.print()
 
+    def get_output_dict(self) -> Dict[str, Any]:
+        """Return dictionary from 4OutputCreator class."""
+        return self.output.file
+
     @staticmethod
     def _check_if_json(input_file: str) -> bool:
         """Check if input file is JSON type."""
@@ -219,7 +201,3 @@ class Solver:
             _LOGGER.warning("File %s is not JSON type. SKIPPED", input_file)
             return False
         return True
-
-    def get_output_dict(self) -> Dict[str, Any]:
-        """Return dictionary from 4OutputCreator class."""
-        return self.output.file
